+++ source/sockets.h
@@ -19,7 +19,7 @@ protected:
 protected:
 	ANT_socket_internals *internals;
 	char *buffer;
-	long buffer_start, buffer_size, buffer_alloc_size, chunk_size;
+	long long buffer_start, buffer_size, buffer_alloc_size, chunk_size;
 	char *poll_getsz_last_pos;
 	unsigned short port;
 	char *gets_result;
@@ -40,7 +40,7 @@ public:
 	int close(void);
 	int error(void);
 
-	char *block_read(char *into, int len);
+	char *block_read(char *into, long long len);
 	char *getsz(char termination = '\0');
 	char *poll_getsz(char termination = '\0');
 	char *gets(void) { return getsz('\n'); }
@@ -48,11 +48,11 @@ public:
 	long getl(void);
 	char getch(void);
 
-	int block_write(const char *buffer, long size);
-	int puts(char *buffer);
-	int putch(char ch);
-	int putsz(char *string);
-	int putl(long what);
+	long long block_write(const char *buffer, long long size);
+	long long puts(char *buffer);
+	long long putch(char ch);
+	long long putsz(char *string);
+	long long putl(long what);
 
 	unsigned int wntohl(unsigned int num);
 	unsigned int whtonl(unsigned int num);
